requirements.txt
python-dotenv==1.0.1 
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8 
unstructured==0.14.4
chromadb==0.5.0 
openai==1.31.1
tiktoken==0.7.0 
nltk 
llama-index
pypdf
langchain-community
streamlit